Refactor: share HBG Graph Execution across architectures
- Move the byte-identical Graph cache, materialization, and host state
  code into a common HBG component.
- Compile the common AICPU source against each architecture's native
  include tree while keeping per-architecture forwarding headers.
- Reuse the shared Graph cache unit source and design document across
  A2/A3 and A5.

Fixes #1737
