+++ src/a2a3/runtime/host_build_graph/build_config.py
@@ -24,7 +24,10 @@
 
 BUILD_CONFIG = {
     "aicore": {"include_dirs": ["runtime", "common", ".."], "source_dirs": ["aicore", "orchestration"]},
-    "aicpu": {"include_dirs": ["runtime", "common", ".."], "source_dirs": ["aicpu", "runtime", "orchestration"]},
+    "aicpu": {
+        "include_dirs": ["runtime", "common", ".."],
+        "source_dirs": ["aicpu", "runtime", "orchestration", "../../../common/host_build_graph"],
+    },
     "host": {
         "include_dirs": ["runtime", "common", ".."],
         "source_dirs": ["host", "runtime/orchestrator_core", "runtime/shared", "orchestration"],
