+++ tests/ut/cpp/CMakeLists.txt
@@ -848,13 +848,13 @@ target_sources(test_a5_hbg_submit_poison PRIVATE
     ${HBG_A5_RUNTIME_DIR}/shared/runtime.cpp
     ${CMAKE_SOURCE_DIR}/../../../src/common/platform/shared/aicpu/args_dump_aicpu.cpp
 )
-add_a2a3_hbg_runtime_test(test_graph_cache a2a3/test_graph_cache.cpp)
+add_a2a3_hbg_runtime_test(test_graph_cache common/test_hbg_graph_cache.cpp)
 target_sources(test_graph_cache PRIVATE
-    ${CMAKE_SOURCE_DIR}/../../../src/a2a3/runtime/host_build_graph/runtime/scheduler/graph_execution.cpp
+    ${CMAKE_SOURCE_DIR}/../../../src/common/host_build_graph/graph_execution.cpp
 )
-add_a5_hbg_runtime_test(test_a5_graph_cache a5/test_graph_cache.cpp)
+add_a5_hbg_runtime_test(test_a5_graph_cache common/test_hbg_graph_cache.cpp)
 target_sources(test_a5_graph_cache PRIVATE
-    ${CMAKE_SOURCE_DIR}/../../../src/a5/runtime/host_build_graph/runtime/scheduler/graph_execution.cpp
+    ${CMAKE_SOURCE_DIR}/../../../src/common/host_build_graph/graph_execution.cpp
 )
 # Incremental-activation tests init a real PTO2SchedulerState (ready queues), so
 # they link the scheduler + shared runtime out-of-line members.
