+++ tests/ut/cpp/common/test_hbg_graph_cache.cpp
@@ -18,6 +18,7 @@
 #include <cstring>
 #include <new>
 #include <thread>
+#include <utility>
 #include <vector>
 
 #include "graph_cache.h"
